dim OLED screen at night
